# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/aging_snapshots:
    get:
      tags:
        - AgingSnapshots
      summary: List AgingSnapshots
      description: Get a paginated list of AgingSnapshots with optional filtering and sorting
      operationId: listAgingSnapshots
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgingSnapshotList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - AgingSnapshots
      summary: Create AgingSnapshot
      description: Create a new AgingSnapshot
      operationId: createAgingSnapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgingSnapshotRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgingSnapshot'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/aging_buckets:
    get:
      tags:
        - AgingBuckets
      summary: List AgingBuckets
      description: Get a paginated list of AgingBuckets with optional filtering and sorting
      operationId: listAgingBuckets
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgingBucketList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - AgingBuckets
      summary: Create AgingBucket
      description: Create a new AgingBucket
      operationId: createAgingBucket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgingBucketRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgingBucket'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    AgingSnapshot:
      type: object
      required:
        - id
        - company_id
        - as_of_date
        - direction
        - total_outstanding
        - bucket_current
        - bucket_1_30
        - bucket_31_60
        - bucket_61_90
        - bucket_90p
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        as_of_date:
          type: string
          format: date
        direction:
          type: string
          maxLength: 10
        total_outstanding:
          type: string
        bucket_current:
          type: string
        bucket_1_30:
          type: string
        bucket_31_60:
          type: string
        bucket_61_90:
          type: string
        bucket_90p:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    CreateAgingSnapshotRequest:
      type: object
      required:
        - company_id
        - as_of_date
        - direction
        - total_outstanding
        - bucket_current
        - bucket_1_30
        - bucket_31_60
        - bucket_61_90
        - bucket_90p
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        as_of_date:
          type: string
          format: date
        direction:
          type: string
          maxLength: 10
        total_outstanding:
          type: string
        bucket_current:
          type: string
        bucket_1_30:
          type: string
        bucket_31_60:
          type: string
        bucket_61_90:
          type: string
        bucket_90p:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    AgingSnapshotList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AgingSnapshot'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    AgingBucket:
      type: object
      required:
        - id
        - company_id
        - snapshot_id
        - invoice_ref
        - invoice_kind
        - due_date
        - days_past_due
        - outstanding_amount
        - bucket
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        snapshot_id:
          type: string
          format: uuid
        invoice_ref:
          type: string
          format: uuid
        invoice_kind:
          type: string
          maxLength: 10
        party_id:
          nullable: true
          type: string
          format: uuid
        due_date:
          type: string
          format: date
        days_past_due:
          type: integer
          format: int32
        outstanding_amount:
          type: string
        bucket:
          type: string
        metadata:
          type: object
...
    CreateAgingBucketRequest:
      type: object
      required:
        - company_id
        - snapshot_id
        - invoice_ref
        - invoice_kind
        - due_date
        - days_past_due
        - outstanding_amount
        - bucket
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        snapshot_id:
          type: string
          format: uuid
        invoice_ref:
          type: string
          format: uuid
        invoice_kind:
          type: string
          maxLength: 10
        party_id:
          nullable: true
          type: string
          format: uuid
        due_date:
          type: string
          format: date
        days_past_due:
          type: integer
          format: int32
        outstanding_amount:
          type: string
        bucket:
          type: string
        metadata:
          type: object
...
    AgingBucketList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AgingBucket'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []