# src/presentation/http/aging_snapshot_handler.rs
//! AgingSnapshot REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use chrono::NaiveDate;
use rust_decimal::Decimal;
use serde::{Deserialize, Serialize};
use uuid::Uuid;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::application::service::{AgingSnapshotService, ServiceError};
use crate::domain::entity::*;

// DTO imports
use crate::presentation::dto::{
    AgingSnapshotResponseDto, CreateAgingSnapshotDto, PatchAgingSnapshotDto, UpdateAgingSnapshotDto,
};

/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum AgingSnapshotError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for AgingSnapshotError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for AgingSnapshotError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "AGINGSNAPSHOT_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "AGINGSNAPSHOT_VALIDATION_ERROR"),
            Self::Database(_) => (
                StatusCode::INTERNAL_SERVER_ERROR,
                "AGINGSNAPSHOT_DATABASE_ERROR",
            ),
            Self::Internal(_) => (
                StatusCode::INTERNAL_SERVER_ERROR,
                "AGINGSNAPSHOT_INTERNAL_ERROR",
            ),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for AgingSnapshot.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /aging_snapshots | List with pagination |
/// | POST | /aging_snapshots | Create new |
/// | GET | /aging_snapshots/:id | Get by ID |
/// | PUT | /aging_snapshots/:id | Full update |
/// | PATCH | /aging_snapshots/:id | Partial update |
/// | DELETE | /aging_snapshots/:id | Soft delete |
/// | POST | /aging_snapshots/bulk | Bulk create |
/// | POST | /aging_snapshots/upsert | Upsert |
/// | GET | /aging_snapshots/trash | List deleted |
/// | POST | /aging_snapshots/:id/restore | Restore |
/// | DELETE | /aging_snapshots/empty | Empty trash |
/// | GET | /aging_snapshots/:id/deleted | Get deleted by ID |
/// | DELETE | /aging_snapshots/trash/:id | Permanent delete from trash |
/// | GET | /aging_snapshots/count | Count active entities |
/// | GET | /aging_snapshots/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(AgingSnapshotService::with_repository(repository));
/// let router = create_aging_snapshot_routes(service);
/// ```
pub fn create_aging_snapshot_routes(service: Arc<AgingSnapshotService>) -> Router {
    BackboneCrudHandler::<
        AgingSnapshotService,
        AgingSnapshot,
        CreateAgingSnapshotDto,
        UpdateAgingSnapshotDto,
        AgingSnapshotResponseDto,
    >::routes(service, "/aging_snapshots")
}

/// Create Axum router with only the read (GET) endpoints for AgingSnapshot.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_aging_snapshot_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_aging_snapshot_read_routes(service: Arc<AgingSnapshotService>) -> Router {
    BackboneCrudHandler::<
        AgingSnapshotService,
        AgingSnapshot,
        CreateAgingSnapshotDto,
        UpdateAgingSnapshotDto,
        AgingSnapshotResponseDto,
    >::read_routes(service, "/aging_snapshots")
}

/// Create Axum router with only the write (mutation) endpoints for AgingSnapshot.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_aging_snapshot_write_routes(service: Arc<AgingSnapshotService>) -> Router {
    BackboneCrudHandler::<
        AgingSnapshotService,
        AgingSnapshot,
        CreateAgingSnapshotDto,
        UpdateAgingSnapshotDto,
        AgingSnapshotResponseDto,
    >::write_routes(service, "/aging_snapshots")
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_aging_snapshot_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<AgingSnapshotService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_aging_snapshot_routes(service).layer(middleware::from_fn(
        move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req
                    .headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| {
                        raw.strip_prefix("Bearer ")
                            .or_else(|| raw.strip_prefix("bearer "))
                    })
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => (
                        axum::http::StatusCode::UNAUTHORIZED,
                        axum::Json(serde_json::json!({
                            "success": false,
                            "error": "unauthorized",
                            "message": "Authentication required"
                        })),
                    )
                        .into_response(),
                }
            }
        },
    ))
}